refactor: split clay.lua into directory structure mirroring suit/
Move from a single 562-line clay.lua to a clay/ directory:
- clay/init.lua: core API, screen composition, layout factory
- clay/tile.lua: tile presets (4 orientations)
- clay/fair.lua: fair presets (2 orientations)
- clay/max.lua: max preset
- clay/corner.lua: corner presets (4 variants)

Each preset file receives the clay module and registers its layouts.
The require path stays the same (Lua resolves clay/init.lua).
